@@ -10,15 +10,17 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "eris": "github:abalabahaha/eris#dev",
+    "@lavacord/discord.js": "^0.0.7",
+    "discord.js": "github:discordjs/discord.js#master",
     "moment": "^2.29.0",
     "moment-duration-format": "^2.3.2"
   },
   "devDependencies": {
     "@types/moment-duration-format": "^2.2.2",
     "@types/node": "^15.0.2",
-    "ts-jest": "^26.3.0",
-    "ts-node-dev": "^1.1.6",
-    "typescript": "^4.0.3"
+    "@types/ws": "^7.4.4",
+    "discord-api-types": "^0.18.1",
+    "dotenv": "^9.0.2",
+    "typescript": "^4.2.4"
   }
 }
