@@ -1,57 +1,21 @@
-import { Client, Shard } from 'eris'
+import { Client, ClientOptions, Collection } from 'discord.js'
 import { readdir } from 'fs'
 
-export default class IllyaClient extends Client {
-  [x: string]: any
+export class IllyaClient extends Client {
   public aliases: any
   public commands: any
-  public shardUptime: any
-  public constructor(token: string, options: object) {
-    super(token, options)
-
-    this.aliases = new Map()
-    this.commands = new Map()
-    this.shardUptime = new Map()
+  constructor(options: ClientOptions) {
+    super(options)
+    this.aliases = new Collection()
+    this.commands = new Collection()
   }
 
-  public connect() {
-    this.shards.forEach((shard: Shard) => {
-      this.shardUptime.set(shard.id, {
-        shardID: shard.id,
-        uptime: 0
-      })
-    })
-    return super.connect()
-  }
-
-  public loadCommands(path: string) {
-    readdir(`${__dirname}/${path}`, (err, f) => {
-      if (err) return console.error(err)
-
-      for (let category of f) {
-        readdir(`${__dirname}/${path}/${category}`, (err, cmd) => {
-          if (err) return console.error(err)
-
-          for (let command of cmd) {
-            const Commands = require(`${__dirname}/${path}/${category}/${command}`).default
-            const commands = new Commands(this)
-            this.commands.set(commands.config.name, commands)
-            commands.config.aliases.forEach((alias: any) => {
-              return this.aliases.set(alias, commands.config.name)
-            })
-          }
-        })
-      }
-    })
-  }
-
-  public loadEvents(path: string) {
-    readdir(`${__dirname}/${path}`, (err, f) => {
-      if (err) return console.error(err)
-
-      f.forEach((files) => {
-        const Events = require(`${__dirname}/${path}/${files}`).default
-        const events = new Events(this)
+  public loadEvents() {
+    readdir(`${__dirname}/events`, (err, files) => {
+      if (err) return console.log(err.message)
+      files.forEach((file) => {
+        const Events = require(`${__dirname}/events/${file}`)
+        const events = new Events[file.split('.')[0]](this)
         super.on(events.name, (...args) => events.run(...args))
       })
     })
