@@ -1,7 +1,5 @@
-import IllyaClient from './Client'
-import config from '../src/config.json'
-const client = new IllyaClient(config.token, config.options)
-
-client.connect()
-client.loadCommands('commands')
-client.loadEvents('events')
+import { IllyaClient } from './Client'
+const config = require('./config')
+const client = new IllyaClient(config.options)
+client.loadEvents()
+client.login(process.env.TOKEN)
